Implement layer sharing via tagging.
The key idea is to annotate sub-trees of a config with the same tags if the
underlying model weights are to be shared. Upon creation of the network, when a
tagged node is encountered, create_child will first look up a global cache for
existing layer with the same tag. If one is found, sanity check is performed to
make sure the network architecture are compatible, then if sanity check is
passed, existing layer is simply returned. If a special tag is first
encountered, a layer with the tag is created and cached.

PiperOrigin-RevId: 459961622