Add a protocol that can be used to override base_hyperparams.sub_config_field (for hparams->fiddle migration).
In particular, FiddleBaseLayer.HParams will be a stub object that implements this protocol.  This will make it possible to migrate an individual layer `L` from being hparams-configured to fiddle-configured, without having to update all instances of `sub_config_field(L.HParams)` (in other layers that use this layer).

Once the migration from HParams->Fiddle is complete, this protocol (and the FiddleBaseLayer.HParams stub object) can be deleted.

PiperOrigin-RevId: 477469991