Generate a more comprehensible and actionable error when we discover that a base class was converted to `FiddleBaseLayer` but one of its subclasses was not.
If we have:

class Parent(FiddleBaseLayer):
  x: int = 0

assert(isinstance(Parent.HParams, _FiddleHParamsClassStub))

And then we try to do:

class Child(Parent):
  class HParams(Parent.HParams):  # [*]
    y: int = 0

Then we'll get the exception on the line marked with "[*]".  In particular, it's trying to create a new subclass of the stub object, which uses the signature of `type.__new__`:

  type(name, bases, dict, **kwds) -> a new type

Without this change, we get a much harder-to-understand error, complaining that __new__ was called with the wrong number of args.

PiperOrigin-RevId: 482261515