Use `pax_fiddle.build` instead of `fdl.build` and `pax_fiddle.Config` instead of `fdl.Config`.
`pax_fiddle.build` is a customized version of `fdl.build` that respects the `DoNotBuild` tag.  If any arguments are tagged with `DoNotBuild`, then they are left as-is (rather than building them).  This makes it possible to keep templates unbuilt, so they can be used for deferred subtree building.

`pax_fiddle.Config` is a customized subclass of `fdl.Config` that adds some compatibility stubs to make it act more like HParams.

PiperOrigin-RevId: 485996015