Fix bug in the way "inheritable" hyperparameters were copied from parent to child for FiddleBaseLayer.
For Fiddle, we were overwriting `child.<foo>` with `parent.<foo>` if `child` didn't have an explicit value for `<foo>` (i.e., if `"<foo>" not in child.__arguments__`).  However, this didn't account for the fact that `child`'s class might overwrite the default value for <foo>.  E.g.:

class MyLayer(FiddleBaseLayer):
  params_init: WeightInit = WeightInit.UniformUnitScaling(0.5)

In this case, even if `child` doesn't have an explicit value for `params_init`, we shouldn't overwrite it with the parent's value.

Updated the copy_base_hparams code to use identical logic for Fiddle and HParams (and not to rely on what's in __arguments__).

Also updated all layers that override params_init to use `pax_fiddle.sub_field`.  This will ensure that the `params_init` fields remain deeply-configurable after we transition to Fiddle.

PiperOrigin-RevId: 487586374