Prevent in-place mutations to `BaseInput.HParams` during instantiation.
With cl/491798831 we start instantiating the actual datasets to trace and get the search space. With this interaction, this means that we end up hardcoding `p.num_infeed_hosts=1` if `=0` originally. This interacts badly with how we configure the num processes and shard index in both `train.py` and `eval_lib.py` since we only override if `=0`, essentially resulting in duplicating data from every process.

PiperOrigin-RevId: 492644144