Pax + Fiddle migration: Rename FiddleBaseLayer and BaseLayerAPI to BaseLayer.
Background: To prepare for the PAX migration from HParams to Fiddle, split the old BaseLayer class into an abstract base class (BaseLayerAPI) and an HParams-specific subclass (that was just named BaseLayer, for backwards compatibility).  We then added a new Fiddle-specific subclass, named FiddleBaseLayer, and migrated all layers from the HParams-specific subclass to the Fiddle subclass.  Finally, we deleted the HParams-specific subclass, merged the Fiddle-specific base class back into its parent, and renamed the resulting class `BaseLayer`.

So at this point, the actual class is named `BaseLayer` and `FiddleBaseLayer` and `BaseLayerAPI` are simply aliases.  This CL replaces references to those aliases with `BaseLayer`.

PiperOrigin-RevId: 494320917