Make sure to use the same dtype on both lhs and rhs throughout quantization.
This CL adds type casting to quantized tensors to make sure both lhs and rhs have the same dtype. lax.dot_general strictly requires both lhs and rhs to have the same dtype. Thus, the weight(rhs) needs to be cast to the dtype of activation(lhs) regardless of its original dtype. The output dtype should follow the activation dtype for the consistent dtype use during fprop. Also, when materializing, the weight and scale in quantize_weight() are cast to int8 and p.dtype, respectively, and then they are materialized for inference.

PiperOrigin-RevId: 494577762