Pax + Fiddle migration: Replace base_layer.sub_config_field(SomeLayer.HParams) with base_layer.template_field(SomeLayer).
All layers have been migrated to use Fiddle, and the `HParams` nested classes are now stub objects.  For these stub objects, `base_layer.sub_config_field(SomeLayer.HParams)` simply returns `base_layer.template_field(SomeLayer)`, so this change should have no impact on the execution of the code.

`base_layer.template_field(SomeLayer)` is a dataclass-style field declaration that indicates that a field should contain a template for `SomeLayer` -- in particular, that it should contain a `pax_fiddle.Config[FiddleBaseLayer]` and that it should default to `pax_fiddle.Config(SomeLayer)`.

This is in contrast to the new `base_layer.instance_field(SomeLayer)`, which can be used to declare fields that directly contain child layers (rather than templates for child layers).

PiperOrigin-RevId: 495138696