Use type signature rather than DoNotBuild tag to decide when not to build fdl.Config objects.
If the type signature for an argument contains `fdl.Config`, then do not build that argument.

Note: for dataclass fields with types like `x: list[Config] = dataclass.field(default_factory=list)`, if the fields is not overridden then it will not be present in the __arguments__ dict, so it will be built by calling the default factory when the constructor is called.  I.e., we will *not* end up with `x` containing `fdl.Config(list)`.
PiperOrigin-RevId: 506391486