Remove cached property from hparams stub properties to avoid deadlocks.
It turns out that the caching was maintaining references to objects which in some cases contained DatasetIterators that caused a deadlock during destruction at program exit in //third_party/py/paxml:seqio_input_test.

Exact details of the deadlock remain unclear to me, it seems to have something to do with the Python interpreter going down before the C++ destructors have finished running (in a different thread?).

In any case, switching from cached_property to plain property fixes the issue.

PiperOrigin-RevId: 513157678