Fix bug in summary type of AGGREGATE_SCALAR, AGGREGATE_IMAGE
This fix updates the summary key type suffixes to preserve the aggregate types, instead of using only the base type in the suffix. However, there remains an ambiguity with the suffixes. This change takes the longest matching type suffix first so that _aggregate_scalar does not match _scalar. Ideally the separator for the suffix '_' would not appear in any of the type names.

PiperOrigin-RevId: 531270934