Fix dimension error when 2D temperature is passed to `sample_decode` and number of samples is 1.
Previously, if a 2D temperature is passed to `sample_decode` with `num_samples == 1` a 3-dimensional temperature is passed to the sample function, leading to an exception like: "IndexError: Too many indices for array: 3 non-None/Ellipsis indices for dim 2."

At line 1091 of `sample_decode.py` the code expects `temperature` to be one-dimensional, so that after adding a new axis it has two.  The function allows two dimensions for the original value of `temperature`, but only flattens it down to 1 before line 1091 if `num_samples > 1`, so this change makes sure it gets flattened when `num_samples == 1` as well.

Included a test case that fails with an exception before this change.

PiperOrigin-RevId: 547539174