Add activation quantization support to `SharedEmbeddingSoftmax.get_logits`.
The input ids to the `emb_lookup` are left unquantized as quantization wouldn't make sense for most values of `SharedEmbeddingSoftmax.num_classes`. The `one_hot_ids` could be cast to `int8` (or `act_params.dtype` if it existed), but I don't know if that's preferable to keeping them in `fprop_dtype`.

PiperOrigin-RevId: 605399430